refactor: remove one level in call stack
