+++ now/cli/__init__.py
@@ -8,8 +8,12 @@ from os.path import expanduser as user
 
 import cpuinfo
 
+from now import __version__
 from now import __version__ as version
+from now.constants import SURVEY_LINK
 from now.deployment.deployment import cmd
+from now.run_all_k8s import start_now, stop_now
+from now.system_information import get_system_state
 
 warnings.filterwarnings("ignore")
 
@@ -117,22 +121,59 @@ def _get_kubectl_path() -> str:
     return kubectl_path
 
 
+def get_task(kwargs):
+    for x in ['cli', 'now']:
+        if x in kwargs:
+            return kwargs[x]
+    raise Exception('kwargs do not contain a task')
+
+
 def cli(args=None):
     """The main entrypoint of the CLI"""
-    if len(sys.argv) != 1 and not ('-h' in sys.argv[1:] or '--help' in sys.argv[1:]):
-        print(f'Initialising Jina NOW v{version} ...')
+    os.environ['JINA_LOG_LEVEL'] = 'CRITICAL'
+    print_version_line()
+    kwargs = parse_args(args)
+    arch, os_type = get_environment_type()
 
+    kwargs['kubectl_path'] = _get_kubectl_path()
+    kwargs['kind_path'] = _get_kind_path()
+
+    contexts, active_context = get_system_state(**kwargs)
+    task = get_task(kwargs)
+    if '--version' in sys.argv[1:]:
+        print(__version__)
+        exit(0)
+    if task == 'start':
+        return start_now(
+            contexts=contexts,
+            active_context=active_context,
+            os_type=os_type,
+            arch=arch,
+            **kwargs,
+        )
+    elif task == 'stop':
+        return stop_now(contexts, active_context, **kwargs)
+    elif task == 'survey':
+        import webbrowser
+
+        webbrowser.open(SURVEY_LINK, new=0, autoraise=True)
+    else:
+        raise Exception(f'unknown task, {task}')
+
+
+def parse_args(args):
     if not args:
         args = _get_run_args()
     args = vars(args)  # Make it a dict from Namespace
+    return args
 
-    if '--version' in sys.argv[1:]:
-        from now import __version__
 
-        print(__version__)
-        exit(0)
+def print_version_line():
+    if len(sys.argv) != 1 and not ('-h' in sys.argv[1:] or '--help' in sys.argv[1:]):
+        print(f'Initialising Jina NOW v{version} ...')
 
-    os.environ['JINA_LOG_LEVEL'] = 'CRITICAL'
+
+def get_environment_type():
     os_type = platform.system().lower()
     arch = 'x86_64'
     if os_type == 'darwin':
@@ -142,12 +183,7 @@ def cli(args=None):
             arch = platform.machine()
     elif os_type == 'linux':
         arch = platform.machine()
-
-    args['kubectl_path'] = _get_kubectl_path()
-    args['kind_path'] = _get_kind_path()
-    from now.run_all_k8s import run_k8s
-
-    return run_k8s(os_type=os_type, arch=arch, **args)
+    return arch, os_type
 
 
 if __name__ == '__main__':
